# docs/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "jobApplier API",
  "version": "0.1.0"
 },
 "paths": {
...
  "/api/profile": {
   "get": {
    "summary": "Read Profile",
    "operationId": "read_profile_api_profile_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   },
   "put": {
    "summary": "Upsert Profile",
    "operationId": "upsert_profile_api_profile_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProfileIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/profile/folder-consent": {
   "post": {
    "summary": "Folder Consent",
    "operationId": "folder_consent_api_profile_folder_consent_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FolderConsentIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Revoke Folder",
    "operationId": "revoke_folder_api_profile_folder_consent_delete",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   }
  },
...
  "/api/documents/import-from-folder": {
   "post": {
    "summary": "Import From Folder",
    "operationId": "import_from_folder_api_documents_import_from_folder_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ImportIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/documents/upload": {
   "post": {
    "summary": "Upload",
    "operationId": "upload_api_documents_upload_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_api_documents_upload_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/documents/linkedin-export": {
   "post": {
    "summary": "Linkedin Export",
    "operationId": "linkedin_export_api_documents_linkedin_export_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_linkedin_export_api_documents_linkedin_export_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/documents/{doc_id}/text": {
   "get": {
    "summary": "Document Text",
    "operationId": "document_text_api_documents__doc_id__text_get",
    "parameters": [
     {
      "name": "doc_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Doc Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/facts": {
   "get": {
    "summary": "List Facts",
    "operationId": "list_facts_api_facts_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   },
   "post": {
    "summary": "Add Fact",
    "description": "User-entered facts: provenance is 'user'. They are still reviewed and approved explicitly.",
    "operationId": "add_fact_api_facts_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FactCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/facts/{fact_id}": {
   "patch": {
    "summary": "Edit Fact",
    "operationId": "edit_fact_api_facts__fact_id__patch",
    "parameters": [
     {
      "name": "fact_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Fact Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FactEdit"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/facts/decide": {
   "post": {
    "summary": "Decide Facts",
    "operationId": "decide_facts_api_facts_decide_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/FactDecision"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/conflicts/{cid}/resolve": {
   "post": {
    "summary": "Resolve Conflict",
    "operationId": "resolve_conflict_api_conflicts__cid__resolve_post",
    "parameters": [
     {
      "name": "cid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cid"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ConflictResolve"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/answers": {
   "get": {
    "summary": "List Answers",
    "operationId": "list_answers_api_answers_get",
    "parameters": [
     {
      "name": "reveal",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Reveal"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/answers/{key}": {
   "put": {
    "summary": "Put Answer",
    "operationId": "put_answer_api_answers__key__put",
    "parameters": [
     {
      "name": "key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AnswerIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "delete": {
    "summary": "Delete Answer",
    "operationId": "delete_answer_api_answers__key__delete",
    "parameters": [
     {
      "name": "key",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Key"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/preferences": {
   "get": {
    "summary": "Read Prefs",
    "operationId": "read_prefs_api_preferences_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     }
    }
   },
   "put": {
    "summary": "Write Prefs",
    "operationId": "write_prefs_api_preferences_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PrefsIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AnswerIn": {
    "properties": {
     "answer": {
      "type": "string",
      "maxLength": 4000,
      "minLength": 1,
      "title": "Answer"
     },
     "approved": {
      "type": "boolean",
      "title": "Approved",
      "default": false
     },
     "sensitive": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sensitive"
     }
    },
    "type": "object",
    "required": [
     "answer"
    ],
    "title": "AnswerIn"
   },
...
   "Body_linkedin_export_api_documents_linkedin_export_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_linkedin_export_api_documents_linkedin_export_post"
   },
   "Body_upload_api_documents_upload_post": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_api_documents_upload_post"
   },
   "ConflictResolve": {
    "properties": {
     "keep": {
      "type": "string",
      "enum": [
       "a",
       "b",
       "both",
       "neither"
      ],
      "title": "Keep"
     },
     "note": {
      "type": "string",
      "maxLength": 500,
      "title": "Note",
      "default": ""
     }
    },
    "type": "object",
    "required": [
     "keep"
    ],
    "title": "ConflictResolve"
   },
...
   "FactCreate": {
    "properties": {
     "kind": {
      "type": "string",
      "enum": [
       "role",
       "achievement",
       "education",
       "skill",
       "certification",
       "project",
       "summary"
      ],
      "title": "Kind"
     },
     "data": {
      "additionalProperties": true,
      "type": "object",
      "title": "Data"
     },
     "parent_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Parent Id"
     }
    },
    "type": "object",
    "required": [
     "kind",
     "data"
    ],
    "title": "FactCreate"
   },
   "FactDecision": {
    "properties": {
     "ids": {
      "items": {
       "type": "integer"
      },
      "type": "array",
      "maxItems": 500,
      "minItems": 1,
      "title": "Ids"
     },
     "decision": {
      "type": "string",
      "enum": [
       "APPROVED",
       "REJECTED",
       "PENDING"
      ],
      "title": "Decision"
     }
    },
    "type": "object",
    "required": [
     "ids",
     "decision"
    ],
    "title": "FactDecision"
   },
   "FactEdit": {
    "properties": {
     "data": {
      "additionalProperties": true,
      "type": "object",
      "title": "Data"
     }
    },
    "type": "object",
    "required": [
     "data"
    ],
    "title": "FactEdit"
   },
   "FolderConsentIn": {
    "properties": {
     "folder": {
      "type": "string",
      "maxLength": 1000,
      "minLength": 1,
      "title": "Folder"
     },
     "consent": {
      "type": "boolean",
      "title": "Consent"
     }
    },
    "type": "object",
    "required": [
     "folder",
     "consent"
    ],
    "title": "FolderConsentIn"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
   "ImportIn": {
    "properties": {
     "name": {
      "type": "string",
      "maxLength": 255,
      "minLength": 1,
      "title": "Name"
     }
    },
    "type": "object",
    "required": [
     "name"
    ],
    "title": "ImportIn"
   },
...
   "ProfileIn": {
    "properties": {
     "first_name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "First Name"
     },
     "last_name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 1,
      "title": "Last Name"
     },
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "phone": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 40
       },
       {
        "type": "null"
       }
      ],
      "title": "Phone"
     },
     "location": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 200
       },
       {
        "type": "null"
       }
      ],
      "title": "Location"
     },
     "timezone": {
      "type": "string",
      "maxLength": 64,
      "title": "Timezone",
      "default": "UTC"
     },
     "linkedin_url": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 2083,
        "minLength": 1,
        "format": "uri"
       },
       {
        "type": "null"
       }
      ],
      "title": "Linkedin Url"
     },
     "portfolio_links": {
      "items": {
       "type": "string",
       "maxLength": 2083,
       "minLength": 1,
       "format": "uri"
      },
      "type": "array",
      "maxItems": 10,
      "title": "Portfolio Links"
     }
    },
    "type": "object",
    "required": [
     "first_name",
     "last_name",
     "email"
    ],
    "title": "ProfileIn"
   },
...
  }
 }
}